refactor(cloud): extract workstation viewer manager
